+++ lib/prompts/snippets/action-types.md
