+++ lib/prompts/snippets/element-types.md
