+++ lib/prompts/snippets/json-output-rules.md
