+++ lib/prompts/templates/interactive-actions/system.md
