+++ lib/prompts/templates/interactive-actions/user.md
