+++ lib/prompts/templates/quiz-actions/system.md
