+++ lib/prompts/templates/quiz-actions/user.md
