+++ lib/prompts/templates/quiz-content/system.md
