+++ lib/prompts/templates/quiz-content/user.md
