+++ lib/prompts/templates/requirements-to-outlines/user.md
