+++ lib/prompts/templates/slide-actions/system.md
