+++ lib/prompts/templates/slide-actions/user.md
