+++ lib/prompts/templates/slide-content/system.md
