+++ lib/prompts/templates/slide-content/user.md
