+++ lib/prompts/templates/web-search-query-rewrite/system.md
