+++ lib/prompts/templates/web-search-query-rewrite/user.md
